backend/pyproject.toml
[project]
name = "ching-tech-os"
version = "0.1.0"
description = "Ching Tech OS Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-socketio>=5.10.0",
    "smbprotocol>=1.12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.0",
    "alembic>=1.13.0",
    "sqlalchemy>=2.0.0",
    "python-multipart>=0.0.20",
    "ptyprocess>=0.7.0",
    "geoip2>=4.8.0",
    "user-agents>=2.2.0",
    "apscheduler>=3.11.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ching_tech_os"]

[tool.hatch.build.targets.sdist]
include = ["src/ching_tech_os"]

[tool.uv]
package = true
